backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
asyncpg==0.29.0
pgvector==0.2.4

# Redis
redis==5.0.1
hiredis==2.3.2

# PDF Processing
PyPDF2==3.0.1
pypdf==3.17.4
python-docx==1.1.0
python-magic==0.4.27

# OpenAI & Embeddings
openai==1.10.0
tiktoken==0.5.2

# Vector DB (FAISS)
faiss-cpu==1.7.4
numpy==1.26.3

# Celery for background jobs
celery==5.3.6
flower==2.0.1

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
aiofiles==23.2.1

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
